Implemented climatology forecast and re-orged utils
