+++ sheerwater_benchmarking/utils/data_utils.py
@@ -3,10 +3,10 @@ import numpy as np
 import xarray as xr
 import xarray_regrid  # noqa: F401, import needed for regridding
 
-from .general_utils import (get_grid_ds,
-                            base360_to_base180, base180_to_base360,
-                            is_wrapped, check_bases,
-                            get_region)
+from .space_utils import (get_grid_ds,
+                          base360_to_base180, base180_to_base360,
+                          is_wrapped, check_bases,
+                          get_region)
 
 
 def apply_mask(ds, mask, var, val=0.0):
@@ -189,7 +189,7 @@ def clip_region(ds, region, lon_dim='lon', lat_dim='lat'):
         lat_dim (str): The name of the latitude dimension.
     """
     region_data = get_region(region)
-    if len(region) == 2:
+    if len(region_data) == 2:
         lon_slice, lat_slice = region_data
     else:
         # Set up dataframe for clipping
