+++ sheerwater_benchmarking/utils/time_utils.py
@@ -1,0 +1,85 @@
+"""Time and date utility functions for all parts of the data pipeline."""
+import dateparser
+from datetime import datetime, timedelta
+from dateutil.rrule import rrule, DAILY, MONTHLY, WEEKLY, YEARLY
+
+
+DATETIME_FORMAT = "%Y-%m-%d"
+
+
+def generate_dates_in_between(start_time, end_time, date_frequency, return_string=False):
+    """Generates dates between two dates based on the frequency.
+
+    Args:
+        start_time (str): The start date.
+        end_time (str): The end date.
+        date_frequency (str): The frequency of the dates.
+            One of "daily", "weekly", a day of the week (e.g., "Monday"), or a combination of days
+            separated by a slash (e.g., "Monday/Thursday").
+        return_string (bool): Whether to return the dates as strings or datetime objects.
+    """
+    start_date = dateparser.parse(start_time)
+    end_date = dateparser.parse(end_time)
+
+    if date_frequency not in ["daily", "weekly"]:
+        dates = [
+            date
+            for date in generate_dates_in_between(start_time, end_time, "daily",
+                                                  return_string=False)
+            if date.strftime("%A") in date_frequency.split("/")
+        ]
+    else:
+        frequency_to_int = {"daily": 1, "weekly": 7}
+        dates = [
+            start_date +
+            timedelta(days=x * frequency_to_int[date_frequency])
+            for x in range(0, int((end_date - start_date).days /
+                                  (frequency_to_int[date_frequency])) + 1,)
+        ]
+
+    if return_string:
+        dates = [date.strftime(DATETIME_FORMAT) for date in dates]
+    return dates
+
+
+def is_valid_forecast_date(model, forecast_type, forecast_date):
+    """Checks if the forecast date is valid for the given model and forecast type."""
+    valid_forecast_dates = {
+        "reforecast": {
+            "ecmwf": ("2015-05-14", datetime.today().strftime(DATETIME_FORMAT), "Monday/Thursday"),
+            "salient": ("2022-01-01", datetime.today().strftime(DATETIME_FORMAT), "Wednesday"),
+        },
+        "forecast": {
+            "ecmwf": ("2015-05-14", datetime.today().strftime(DATETIME_FORMAT), "Monday/Thursday"),
+            "salient": ("2022-01-01", datetime.today().strftime(DATETIME_FORMAT), "Wednesday"),
+        },
+    }
+    assert isinstance(forecast_date, datetime)
+    try:
+        return forecast_date in generate_dates_in_between(
+            *valid_forecast_dates[forecast_type][model], return_string=False)
+    except KeyError:
+        return False
+
+
+def get_dates(start_time, end_time, stride="day", return_string=True):
+    """Outputs the list of dates corresponding to input date string."""
+    # Input is of the form '20170101-20180130'
+    start_date = dateparser.parse(start_time)
+    end_date = dateparser.parse(end_time)
+
+    if stride == "day":
+        stride = DAILY
+    elif stride == "week":
+        stride = WEEKLY
+    elif stride == "month":
+        stride = MONTHLY
+    elif stride == "year":
+        stride = YEARLY
+    else:
+        raise ValueError(
+            "Only day, week, month, and year strides are supported.")
+    dates = [dt for dt in rrule(stride, dtstart=start_date, until=end_date)]
+    if return_string:
+        dates = [date.strftime(DATETIME_FORMAT) for date in dates]
+    return dates
